Split consesus.rs
Attempt to split consensus.rs from sandbox into meaningful
pieces.

basic.rs - basic consensus tests
round_detail.rs - more elaborate consensus tests
block_request.rs - block request related tests
config.rs - config change related tests
invalid_message.rs - tests related to malformed messages
recovery.rs - tests related to the recovery of the node after restart
transactions.rs - transactions handling
unsynchronized_message.rs - tests related to messages that arrive
at the wrong time.
